+++ utils/src/main/lombok/de/larssh/utils/xml/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Utilities to process XML documents
+ */
+@de.larssh.utils.annotations.NonNullByDefault
+package de.larssh.utils.xml;