+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Download at [Maven Search](https://search.maven.org/search?q=g:de.lars-sh%20AND%
 
 ### Added
 * Checkstyle check `XAvoidSerialVersionUID`, warning on usage of `serialVersionUID`.
+* Constants for Xerces features (see `de.larssh.utils.xml.XercesFeatures`)
 
 ### Changed
 * [Avoid using serialVersionUID](https://github.com/rzwitserloot/lombok/wiki/WHY-NOT:-serialVersionUID).